libfuse: pass security context options to kernel
Mount can be used with an "-o context=" option in order to specify a
mountpoint-wide SELinux security context different from the default
context
provided by the active SELinux policy.

This is useful in order to enable users to mount multiple sshfs targets
under
distinct contexts, which is my main motivation for getting this patch
mainlined.

Change-Id: If75cd36828afe0e56cd8c8dfc41bdbf77f12e206